REFACTOR: 불필요한 파일 삭제 및 DTO 디렉토리 구조 개선
